move instances utils to folder
