+++ hat/assets/js/apps/Iaso/domains/instances/utils/index.js
@@ -7,19 +7,19 @@ import { truncateText, getTableUrl } from 'bluesquare-components';
 import { FormattedMessage } from 'react-intl';
 import CompareArrowsIcon from '@material-ui/icons/CompareArrows';
 
-import instancesTableColumns, { actionTableColumn } from './config';
-import MESSAGES from './messages';
-import DeleteDialog from './components/DeleteInstanceDialog';
-import ExportInstancesDialogComponent from './components/ExportInstancesDialogComponent';
-import { getCookie } from '../../utils/cookies';
+import instancesTableColumns, { actionTableColumn } from '../config';
+import MESSAGES from '../messages';
+import DeleteDialog from '../components/DeleteInstanceDialog';
+import ExportInstancesDialogComponent from '../components/ExportInstancesDialogComponent';
+import { getCookie } from '../../../utils/cookies';
 import {
     apiDateTimeFormat,
     apiDateFormat,
     getFromDateString,
     getToDateString,
-} from '../../utils/dates.ts';
-import { fetchLatestOrgUnitLevelId } from '../orgUnits/utils';
-import { baseUrls } from '../../constants/urls';
+} from '../../../utils/dates.ts';
+import { fetchLatestOrgUnitLevelId } from '../../orgUnits/utils';
+import { baseUrls } from '../../../constants/urls';
 
 const NO_VALUE = '/';
 const hasNoValue = value => !value || value === '';
